requirements.txt
# 執行依賴（包進 .exe 裡）
PySide6==6.7.3
mss==9.0.2
Pillow==10.4.0
numpy>=1.24,<2.4
opencv-contrib-python==4.10.0.84
pywinctl==0.4.1
paddlepaddle==3.3.0
paddleocr==3.5.0
